+++ FBSnapshotsViewer/Managers/SnapshotTestResultSwapper.swift
@@ -11,6 +11,7 @@ import Nuke
 
 enum SnapshotTestResultSwapperError: Error {
     case canNotBeSwapped(testResult: SnapshotTestResult)
+    case canNotBeRejected(testResult: SnapshotTestResult)
     case nonRetinaImages(testResult: SnapshotTestResult)
     case notExistedRecordedImage(testResult: SnapshotTestResult)
     case canNotPerformFileManagerOperation(testResult: SnapshotTestResult, underlyingError: Error)
@@ -61,6 +62,20 @@ class SnapshotTestResultSwapper {
             try fileManager.moveItem(at: failedImageURL, to: recordedImageURL)
             imageCache.invalidate()
             return SnapshotTestResult.recorded(testInformation: testInformation, referenceImagePath: recordedImageURL.path, build: build)
+        }
+        catch let error {
+            throw SnapshotTestResultSwapperError.canNotPerformFileManagerOperation(testResult: testResult, underlyingError: error)
+        }
+    }
+
+    func reject(_ testResult: SnapshotTestResult) throws -> SnapshotTestResult {
+        guard case let SnapshotTestResult.failed(testInformation, referenceImagePath, _, _, build) = testResult, canSwap(testResult) else {
+            throw SnapshotTestResultSwapperError.canNotBeRejected(testResult: testResult)
+        }
+        do {
+            try removeTestImages(testResult)
+            imageCache.invalidate()
+            return SnapshotTestResult.rejected(testInformation: testInformation, referenceImagePath: referenceImagePath, build: build)
         }
         catch let error {
             throw SnapshotTestResultSwapperError.canNotPerformFileManagerOperation(testResult: testResult, underlyingError: error)